refactor vault little bit
